Renamed SettingsBar to ListHeader removed un-used imports on App.js
