+++ src/components/ListHeader/ListHeader.js
@@ -1,4 +1,4 @@
-import React, { PureComponent } from 'react'
+import React from 'react'
 import { connect } from 'react-redux'
 import { View } from 'react-native'
 import PropTypes from 'prop-types'
@@ -8,16 +8,16 @@ import ButtonMenu from '../ButtonMenu/ButtonMenu'
 import { toggleNav } from '../../state/actions'
 
 import ustyle from '../../utils/style'
-import style from './SettingsBar.style.js'
+import style from './ListHeader.style.js'
 
-const SettingsBar = ({ list, isNavOpen, toggleNav }) =>
-  <View style={[style.view, ustyle.fr]}>
-    { list && <ButtonMenu list={list}/> }
-    <NavButton onPress={toggleNav} isNavOpen={isNavOpen} />
-  </View>
+const ListHeader = ({ list, isNavOpen, toggleNav }) =>
+    <View style={[style.view, ustyle.fr]}>
+      { list && <ButtonMenu list={list}/> }
+      <NavButton onPress={toggleNav} isNavOpen={isNavOpen} />
+    </View>
 
 
-SettingsBar.propTypes = {
+ListHeader.propTypes = {
   list: PropTypes.object,
   isNavOpen : PropTypes.bool,
 }
@@ -26,4 +26,4 @@ const mapDispatchToProps = {
   toggleNav,
 }
 
-export default connect(null, mapDispatchToProps)(SettingsBar)
+export default connect(null, mapDispatchToProps)(ListHeader)
